+++ custom_components/unifi_insights/api/network/models/__init__.py
@@ -67,6 +67,7 @@ from .traffic import (
     TrafficMatchingType,
 )
 from .voucher import Voucher, VoucherCreateRequest
+from .vpn_client import VpnClient
 from .wifi import WifiNetwork, WifiSecurity
 
 __all__ = [
@@ -132,6 +133,7 @@ __all__ = [
     "VPNTunnelStatus",
     "WANInterface",
     "WANStatus",
+    # Routes
     # Site
     "Site",
     "SiteHealth",
@@ -144,6 +146,8 @@ __all__ = [
     # Voucher
     "Voucher",
     "VoucherCreateRequest",
+    # VPN Client
+    "VpnClient",
     # WiFi
     "WifiNetwork",
     "WifiSecurity",